Trait tvm::function::Typed

pub trait Typed<I, O> {
    fn args(i: Vec<ArgValue<'static>>) -> Result<I, Error>;

    fn ret(o: O) -> Result<RetValue, Error>;
}

A trait representing whether the function arguments and return type can be assigned to a TVM packed function.

By splitting the conversion to function into two traits we are able to improve error reporting, by splitting the conversion of inputs and outputs to this trait.

And the implementation of it to ToFunction.

Required methods

fn args(i: Vec<ArgValue<'static>>) -> Result<I, Error>

fn ret(o: O) -> Result<RetValue, Error>

Loading content...