Module tvm::runtime::function

This module provides an idiomatic Rust API for creating and working with TVM functions.

For calling an already registered TVM function use [function::Builder] To register a TVM packed function from Rust side either use [function::register] or the macro [register_global_func].

See the tests and examples repository for more examples.

Modules

ffi

The low-level C runtime FFI API for TVM.

Structs

Function

An owned thread-safe version of tvm::PackedFunc for consumption in Rust.

RawArgs

Enums

ArgValue

A borrowed TVMPODValue. Can be constructed using into() but the preferred way to obtain a ArgValue is automatically via call_packed!.

RetValue

An owned TVMPODValue. Can be converted from a variety of primitive and object types. Can be downcasted using try_from if it contains the desired type.

Traits

ToFunction

Typed

Functions

register

Registers a Rust function with an arbitrary type signature in the TVM registry.

register_override

Register a function with explicit control over whether to override an existing registration or not.

register_untyped

Type Definitions

Result