Crate tvm

TVM is a compiler stack for deep learning systems.

This crate provides an idiomatic Rust API for TVM runtime frontend.

One particular use case is that given optimized deep learning model artifacts, (compiled with TVM) which include a shared library lib.so, graph.json and a byte-array param.params, one can load them in Rust idiomatically to create a TVM Graph Runtime and run the model for some inputs and get the desired predictions all in Rust.

Checkout the examples repository for more details.

Modules

context

errors

function
This module provides an idiomatic Rust API for creating and working with TVM functions.

ir

module
Provides the [Module] type and methods for working with runtime TVM modules.

ndarray
This module implements the [NDArray] type for working with TVM tensors or coverting from a Rust's ndarray to TVM NDArray.

python

runtime

transform

value
This module implements [ArgValue] and [RetValue] types and their conversions needed for the types used in frontend crate. RetValue is the owned version of TVMPODValue.

Macros

export

export_mod

export_pass

initialize
A macro for generating the correct TVM symbols for plugin loading.

Structs

Context

DataType

Function
Wrapper around TVM function handle which includes is_global indicating whether the function is global or not, and is_cloned showing not to drop a cloned function from Rust side. The value of these fields can be accessed through their respective methods.

FunctionNotFoundError

Module
Wrapper around TVM module handle which contains an entry function. The entry function can be applied to an imported module through entry_func.

NDArray

TypeMismatchError

Enums

DeviceType
Device type represents the set of devices supported by TVM.

Error

NDArrayError

Functions

version
Outputs the current TVM version.