Crate tvm

TVM is a compiler stack for deep learning systems.

This crate provides an idiomatic Rust API for TVM runtime frontend.

One particular use case is that given optimized deep learning model artifacts, (compiled with TVM) which include a shared library lib.so, graph.json and a byte-array param.params, one can load them in Rust idiomatically to create a TVM Graph Executor and run the model for some inputs and get the desired predictions all in Rust.

Checkout the examples repository for more details.

Modules

compiler

device

errors

function

This module provides an idiomatic Rust API for creating and working with TVM functions.

ir

module

Provides the Module type and methods for working with runtime TVM modules.

ndarray

This module implements the NDArray type for working with TVM tensors or coverting from a Rust’s ndarray to TVM NDArray.

python

runtime

transform

Macros

export

export_mod

export_pass

initialize

A macro for generating the correct TVM symbols for plugin loading.

Structs

DataType

Device

Function

An owned thread-safe version of tvm::PackedFunc for consumption in Rust.

FunctionNotFoundError

Module

NDArray

TypeMismatchError

Enums

DeviceType

Device type represents the set of devices supported by TVM.

Error

NDArrayError

Functions

version

Outputs the current TVM version.