Trait tvm::runtime::object::IsObject

pub unsafe trait IsObject: AsRef<Object> + Debug {
    const TYPE_KEY: &'static str;
    unsafe extern "C" fn typed_delete(object: *mut Self) { ... }
}

An unsafe trait which should be implemented for an object subtype.

The trait contains the type key needed to compute the type index, a method for accessing the base object given the subtype, and a typed delete method which is specialized to the subtype.

Associated Constants

const TYPE_KEY: &'static str

Loading content...
Provided methods

unsafe extern "C" fn typed_delete(object: *mut Self)

Loading content...